use bcsv key for rocksdb storage path
