+++ Cargo.toml
@@ -26,6 +26,7 @@ codec = { package = "parity-scale-codec", version = "3.0.0", features = [
 color-print = "0.3.4"
 futures-timer = "3.0.1"
 hash-db = "0.16.0"
+hex = "0.4.3"
 hex-literal = { version = "0.4.1" }
 jsonrpsee = { version = "0.22", features = ["server"] }
 kvdb = "0.13.0"
