+++ client/forest-manager/src/prove.rs
@@ -81,78 +81,12 @@ pub(crate) fn prove<T: TrieLayout, F: ForestStorage>(
 
 #[cfg(test)]
 mod tests {
-    use crate::in_memory::InMemoryForestStorage;
-
-    use common::types::HashT;
+    use crate::{in_memory::InMemoryForestStorage, tests::build_merkle_patricia_forest};
 
     use super::*;
     use reference_trie::RefHasher;
-    use sp_core::H256;
-    use sp_trie::{LayoutV1, MemoryDB};
-    use storage_hub_infra::types::Metadata;
-    use trie_db::{Hasher, TrieDBBuilder, TrieDBMutBuilder, TrieMut};
-
-    /// Build a Merkle Patricia Forest Trie.
-    ///
-    /// The trie is built from the ground up, by each file into 32 byte chunks and storing them in a trie.
-    /// Each trie is then inserted into a new merkle patricia trie, which comprises the merkle forest.
-    pub fn build_merkle_patricia_forest<T: TrieLayout>() -> (
-        MemoryDB<T::Hash>,
-        HashT<T>,
-        Vec<<<T as TrieLayout>::Hash as Hasher>::Out>,
-    ) {
-        let user_ids = vec![
-            b"01", b"02", b"03", b"04", b"05", b"06", b"07", b"08", b"09", b"10", b"11", b"12",
-            b"13", b"12", b"13", b"14", b"15", b"16", b"17", b"18", b"19", b"20", b"21", b"22",
-            b"23", b"24", b"25", b"26", b"27", b"28", b"29", b"30", b"31", b"32",
-        ];
-        let bucket = b"bucket";
-        let file_name = b"sample64b";
-
-        let mut file_leaves = Vec::new();
-
-        for user_id in user_ids {
-            let file_path = format!(
-                "{}-{}-{}.txt",
-                String::from_utf8(user_id.to_vec()).unwrap(),
-                String::from_utf8(bucket.to_vec()).unwrap(),
-                String::from_utf8(file_name.to_vec()).unwrap()
-            );
-
-            let fingerprint = H256::from_slice(&[0; 32]);
-
-            let metadata = Metadata {
-                owner: String::from("owner"),
-                location: file_path.as_bytes().to_vec(),
-                size: 0,
-                fingerprint,
-            };
-
-            let metadata = bincode::serialize(&metadata).unwrap();
-            let metadata_hash = T::Hash::hash(&metadata);
-
-            file_leaves.push((metadata_hash, metadata));
-        }
-
-        // Construct the Merkle Patricia Forest
-        let mut memdb = MemoryDB::<T::Hash>::default();
-        let mut root: HashT<T> = Default::default();
-
-        let mut file_keys = Vec::new();
-        {
-            let mut merkle_forest_trie = TrieDBMutBuilder::<T>::new(&mut memdb, &mut root).build();
-
-            // Insert file leaf and metadata into the Merkle Patricia Forest.
-            for file in &file_leaves {
-                merkle_forest_trie
-                    .insert(file.0.as_ref(), file.1.as_ref())
-                    .unwrap();
-
-                file_keys.push(file.0);
-            }
-        }
-        (memdb, root, file_keys)
-    }
+    use sp_trie::LayoutV1;
+    use trie_db::TrieDBBuilder;
 
     #[test]
     fn test_prove_challenge_exact_key_match() {
